requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
decorator==4.4.2
Django==4.1.7
djangorestframework==3.14.0
ez-setup==0.9
ffmpeg==1.4
idna==3.4
imageio==2.26.1
imageio-ffmpeg==0.4.8
numpy==1.24.2
Pillow==9.4.0
proglog==0.1.10
pytz==2023.2
requests==2.31.0
sqlparse==0.4.4
tqdm==4.65.0
tzdata==2023.2
urllib3==1.26.15
django-embed-video==1.4.9
